+++ src/lib/typeutils.ts
@@ -1,0 +1,1 @@
+export type Maybe<T> = T | null
